better fuzz test code sharing
