+++ angular_analyzer_plugin/test/fuzzing/case.dart
@@ -1,0 +1,7 @@
+class FuzzCase {
+  final int transformCount;
+  final String dart;
+  final String html;
+
+  FuzzCase(this.transformCount, this.dart, this.html);
+}
